+++ CHANGES.md
@@ -1,6 +1,12 @@
 twilio-node changelog
 =====================
 
+[2017-05-24] Version 3.2.0-alpha-1
+----------------------------------
+- Add HostedNumbers preview support.
+- Add Proxy preview support.
+- Add BulkExports preview support.
+
 [2017-05-03] Version 3.1.0-alpha-1
 ----------------------------------
  - Add video domain.